==> Scripts/7-NYC-SourcesMerged.ipynb <==
# NYC Rodent Merged Dataset
### 2018-2023

#### _Overview_ 
This notebook merges aggregated datasets on rodent sightings, sanitation complaints, subways (todo) and restaurant inspections (todo).

from datetime import datetime, timedelta
import geopandas as gpd
import json
import pandas as pd
import mapclassify
import matplotlib.pyplot as plt
import numpy as np
import os
import requests
from io import StringIO
import warnings

warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)
pd.set_option("display.max_rows", None)
np.set_printoptions(threshold=np.inf)

# Parameters
parent_dir = os.path.abspath('..')  # get the absolute path of the parent directory

## Read Aggregated (cleaned) Data Sources

# Read Census Block Groups
cb_file_path = os.path.join(parent_dir, 'Data', 'nyc_bgrp.geojson')  # construct the file path
cb_gdf = gpd.read_file(cb_file_path)  # load the GeoJSON file into a GeoDataFrame
print(cb_gdf.dtypes)
print(len(cb_gdf))
cb_gdf.head(2)


# Read Rodent Sightings (311)
sightings_file_path = os.path.join(parent_dir, 'Data', 'sightings_per_year.csv')  # construct the file path
sightings_df = pd.read_csv(sightings_file_path)
sightings_df['spatial_id'] = sightings_df['spatial_id'].astype(str)
print(len(sightings_df))
print(sightings_df.dtypes)
sightings_df.head(2)


# Read DSNY Complaints
dsny_complaints_file_path = os.path.join(parent_dir, 'Data', 'dsny_complaints_per_year.csv')  # construct the file path
dsny_complaints_df = pd.read_csv(dsny_complaints_file_path)
dsny_complaints_df['spatial_id'] = dsny_complaints_df['spatial_id'].astype(str)
print(len(dsny_complaints_df))
print(dsny_complaints_df.dtypes)
dsny_complaints_df.head(2)


# Read Subway
subway_file_path = os.path.join(parent_dir, 'Data', 'subway_entrances_2_censusblocks.csv')  # construct the file path
subway_df = pd.read_csv(subway_file_path)
subway_df['spatial_id'] = subway_df['spatial_id'].astype(str)
print(len(subway_df))
print(subway_df.dtypes)
subway_df.head(2)


# Read Restaurant Violations
rest_violations_file_path = os.path.join(parent_dir, 'Data', 'violations_per_year.csv')  # construct the file path
rest_violations_df = pd.read_csv(rest_violations_file_path)
rest_violations_df['spatial_id'] = rest_violations_df['spatial_id'].astype(str)
print(len(rest_violations_df))
print(rest_violations_df.dtypes)
rest_violations_df.head(2)

## Merge Data

cbg_merged_gdf = pd.merge(cb_gdf, sightings_df, on=['spatial_id'], how='left')
cbg_merged_gdf = pd.merge(cbg_merged_gdf, dsny_complaints_df, on=['spatial_id','year'], how='left')
cbg_merged_gdf = pd.merge(cbg_merged_gdf, subway_df, on=['spatial_id'], how='left')
cbg_merged_gdf = pd.merge(cbg_merged_gdf, rest_violations_df, on=['spatial_id','year'], how='left')

print(len(cbg_merged_gdf))
cbg_merged_gdf.head(2)

## Normalize columns by Population

# TODO 

## Save Merged Dataset

export_columns = ['spatial_id'] + cbg_merged_gdf.columns[7:].to_list() + ['MedianHouseholdIncome']
print(export_columns)
parent_dir = os.path.abspath('..')  # get the absolute path of the parent directory
file_path = os.path.join(parent_dir, 'Data', 'rodents_per_year_merged.csv')  # construct the file path
cbg_merged_gdf[export_columns].to_csv(file_path, index=False)
