==> Scripts/8-PCA.ipynb <==
# Principal Component Analysis (EDA)

This notebook applies Principal Component Analysis (PCA) to the merged rodent_df dataset. The output are the principal components and their explained variance, plus the importance score of the top 10 columns. This information can be used to identify the most important variables in the data, which can help with feature selection and dimensionality reduction.

It does not include certain fields in the analysis such as start with 't_' or 'd_' (breakdowns of the rodent sightings per time). It then drops three columns, 'year', 'num_dsny_complaints', and 'spatial_id'.

from datetime import datetime, timedelta
import geopandas as gpd
import json
import pandas as pd
import mapclassify
import matplotlib.pyplot as plt
import numpy as np
import os
import requests
from sklearn.decomposition import PCA
from io import StringIO
import warnings

warnings.filterwarnings('ignore')

pd.set_option('display.max_columns', None)
pd.set_option("display.max_rows", None)
np.set_printoptions(threshold=np.inf)

## Load Data

# Parameters
parent_dir = os.path.abspath('..')  # get the absolute path of the parent directory

file_path = os.path.join(parent_dir, 'Data', 'rodents_per_year_merged.csv')  # construct the file path
rodents_df = pd.read_csv(file_path)
print(len(rodents_df))
rodents_df.head()

# Run PCA

# Load your high-dimensional data into a pandas DataFrame
df = rodents_df.filter(regex='^(?!t_|d_).*')
df = df.drop(['year', 'num_sightings', 'num_dsny_complaints','spatial_id'], axis=1)
#print(df.dtypes)

# Center and scale the data
X = df.values
X = (X - np.mean(X, axis=0)) / np.std(X, axis=0)

# Create a PCA object and fit it to the data
pca = PCA()
pca.fit(X)

# Access the principal components and their explained variance
pcs = pca.components_
variance = pca.explained_variance_ratio_

# Print the explained variance for each principal component
for i, var in enumerate(variance):
    print("PC", i+1, "explains", round(var*100, 2), "% of the variance")

# Determine the most important columns
importance_scores = np.abs(pcs)
column_importance = pd.DataFrame(importance_scores.T, index=df.columns)

important_cols = []

# Print the top 10 columns and their importance for the first 5 components
for i in range(5):
    print("\nTop 10 columns for PC", i+1)
    importance = column_importance[i].nlargest(10)
    top_10 = column_importance[i].nlargest(10)
    top_10_fields = list(top_10.index)
    important_cols += top_10_fields
    print(top_10_fields)
    print(importance)

print(list(set(important_cols)))
